Moved git related tasks and plugins to appropriate package
