+++ src/main/groovy/org/shipkit/internal/gradle/VersioningPlugin.java
@@ -6,6 +6,7 @@ import org.gradle.api.Project;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
 import org.shipkit.gradle.BumpVersionFileTask;
+import org.shipkit.internal.gradle.git.GitPlugin;
 import org.shipkit.internal.gradle.util.TaskMaker;
 import org.shipkit.internal.version.Version;
 import org.shipkit.internal.version.VersionInfo;
