+++ src/main/groovy/org/shipkit/internal/gradle/release/CiReleasePlugin.java
@@ -8,8 +8,8 @@ import org.gradle.api.logging.Logging;
 import org.gradle.api.tasks.StopExecutionException;
 import org.gradle.process.ExecResult;
 import org.shipkit.gradle.exec.ShipkitExecTask;
-import org.shipkit.internal.gradle.GitSetupPlugin;
 import org.shipkit.internal.gradle.ReleaseNeededPlugin;
+import org.shipkit.internal.gradle.git.GitSetupPlugin;
 import org.shipkit.internal.gradle.util.TaskMaker;
 
 import static java.util.Arrays.asList;
