+++ src/test/groovy/org/shipkit/internal/gradle/git/GitPluginTest.groovy
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle
+package org.shipkit.internal.gradle.git
 
 import testutil.PluginSpecification
